Implemented that a user can login
In detail i made the LoginValidator.java class that does the validation for the login. Updated the onLogin methood in LoginController.java so that a user only can log in if it is registered. Made userInfoFinder.java which as of now finds the fullname og a user based on its username.