Updated writing to and reading from json-file
Updated FileOperator.java to use jackson to read from and write to file. Made a new class DataObjects.java that represents the dataObjects.json-file as a java.util.hashMap. Had to update the controllers, logInValidator and RegoistrationValidator to match the chnages in FileOperator. Found a double referance in User.java and Ad.java- classes and removed it.