Cleaned up the pom-files
The pom-structure is now good. There is no duplicate dependencies and the child poms is using the parent pom in the correcr way. Also added link a for all the dependencies to specify where the dependencies is retrived from. Had to fix one spotbug in remoteAccsess to be sure that the new structure is good.