Implemented jpackage
Added Jlink and Jpackage to the pom.xml file in ui module. It is now possible to export the prosject as an application. Had to remove org.json dependecy so that Jlink would work. org.json has auto generated module-info and Jlink doesnt work with dependencies that has auto generated module-infos. Because of this i removed the dependency and updateted Dataobjects.java to work without using org.json.